chromebook.el: wrap our changes up in a global minor-mode
This makes it easier to over-ride the key-sequences for quitting the
window. I've set it up so the mode is only entered if the new frame is
on the X display. As you have to exit the X frame before you could enter
anything on a tty window the fact there is a global mode shouldn't be a
problem.