chromebook.el: disable FBC on Intel hardware
This is a workaround to a bug with Frame Buffer compression which was
recently re-enabled on ChromeOS. Although it works fine for the browser
it causes problems for host-x11 apps like Emacs and causes corruption on
the screen.

This patch disabled FBC on the new frame although the effect doesn't
seem to kick in into next iteration. As this has a deleterious effect on
battery performance it should ideally be disabled when we are no longer
displaying an Emacs frame.